+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+- Fix a bug where API hangs when the user has zero backups
+
 ## Release 1.0.0 (2025-02-10)
 
 - Upgrade NPM dependencies